consolidated graph/keyword/entry load into own component, removed from graph.js. prepping to setup twitter page
